refactor: consolidate git keyword classes with base class
Extract shared behavior into Git::Base class. Each keyword class
now only defines its OPTIONS constant. Consolidate tests into
single file with command verification for each keyword.

Also add keyword methods to Arguments::Files so keywords translate
to file lists (unstaged, modified, untracked).

Note: File lists are collected but not yet passed to tools - that
requires completing the File Targeting feature.

- Reduces ~140 lines to ~53 lines of implementation
- Reduces ~170 lines to ~63 lines of tests
- Adds command string tests to catch OPTIONS typos
